+++ src/TestSuite/Traits/MockTrait.php
@@ -18,6 +18,7 @@ use Cake\Controller\Controller;
 use Cake\Datasource\ConnectionManager;
 use Cake\ORM\Entity;
 use Cake\Utility\Inflector;
+use Cake\View\View;
 
 /**
  * A mock trait
@@ -68,6 +69,33 @@ trait MockTrait
 
         return $this->getMockBuilder($className)
             ->setConstructorArgs([null, null, $alias])
+            ->setMethods($methods)
+            ->getMock();
+    }
+
+    /**
+     * Mocks an helper
+     * @param string $className Helper class name
+     * @param array|null $methods The list of methods to mock
+     * @return \PHPUnit\Framework\MockObject\MockObject
+     */
+    protected function getMockForHelper($className, $methods = [])
+    {
+        return $this->getMockBuilder($className)
+            ->setMethods($methods)
+            ->setConstructorArgs([new View])
+            ->getMock();
+    }
+
+    /**
+     * Mocks a shell
+     * @param string $className Shell class name
+     * @param array|null $methods The list of methods to mock
+     * @return \PHPUnit\Framework\MockObject\MockObject
+     */
+    protected function getMockForShell($className, $methods = ['_stop', 'in'])
+    {
+        return $this->getMockBuilder($className)
             ->setMethods($methods)
             ->getMock();
     }