+++ CHANGELOG.md
@@ -6,7 +6,10 @@
 
 ## 2.17 branch
 ### 2.17.5
-* added `MockTrait` and `ComponentTestCase` classes for test suite.
+* added `MockTrait`, `ComponentTestCase` and `HelperTestCase` classes for test
+    suite;
+* added `getParserDescription()`, `getParserOptions()` and
+    `getParserSubcommands()` methods for the `ConsoleIntegrationTestCase` class.
 
 ### 2.17.4
 * fixed bug in the integration with CKeditor when uploading images (see